Working
Now works with android Smartphone alone or if you deicde to use a Smartwatch.

IMPORTANT: Smart watch functionally only works first run. If die and tap to play again it will not work on second run and app crashes. BUT, if die and tap to play again it works on the 3rd run. It will always work on an odd number of runs and will crash on an even number of runs.